Working recursive version!!!!
Using a modified minimax type of pruning we know have a successful
recursive approach to the problem with a heuristic positional judgement
after every move that maximizes for the computer and minimizes for the
opponent. Game play by the computer is currently reasonably smart. Need
to make adjustments to heuristics. Also made a neat printing system for
the recursion by using tabs.

The computers moves that it makes while thinking are currently visible
for the world to see. I might make a fix, and then comment it out,
because not gonna lie it looks pretty cool.