Computer Beats Me!
Completely working. Freebies are successful. The computer beat me and I
tried 100% (DEPTHLIMIT was 5 moves, so it was making its moves in under
a second). I'm very happy (good commit message content...). Maybe there
can be some refinement to the values for heuristics but it seems like I
chose well.

We still need to make different difficulty levels. And maybe see if a
brute force solution can be done to make myself sad.