Bug fixes and clean up
Fully working computer now. Need to add change in DEPTHLIMIT based
on the depth of the game. But now it beats me on Medium (which is a
moderate think tank). Hard is probably stronger, but needs some pruning
added to it.

Changes include:
-Added a computer check to see if the game was done (so that it didn't
go further and totally mis evaluate the position.
-Made the next big box correct when being sent to a taken box (for the
computers evaluations)

We had added a whole tree structure to keep from having to re-evaluate
the position every time, but because we only stored what we had seen, by
the time we reach the next computer run through, the trees usefulness
had already been reduced by a factor of 81. So that was totally scraped
without even a commit to its name.

Future plans:
We need to consider making the computer think during the other
players turn (and if so, make it breadth first at that point).