Fixed bugs
Completely working fixes. I thought the last commit would solve my
issues but it turned out that it didn't, so this one was needed. Now
there is no playing in boxes that have been taken, which was solved by
not using the bigstates array, but rather the individual values of the
board.big.state values.